+++ mindee/input/sources/__init__.py
@@ -1,0 +1,6 @@
+from mindee.input.sources.base_64_input import Base64Input
+from mindee.input.sources.bytes_input import BytesInput
+from mindee.input.sources.file_input import FileInput
+from mindee.input.sources.local_input_source import InputType, LocalInputSource
+from mindee.input.sources.path_input import PathInput
+from mindee.input.sources.url_input_source import UrlInputSource
