+++ mindee/input/sources/bytes_input.py
@@ -1,0 +1,19 @@
+import io
+
+from mindee.input import InputType, LocalInputSource
+
+
+class BytesInput(LocalInputSource):
+    """Raw bytes input."""
+
+    def __init__(self, raw_bytes: bytes, filename: str) -> None:
+        """
+        Input document from raw bytes (no buffer).
+
+        :param raw_bytes: Raw data as bytes
+        :param filename: File name of the input
+        """
+        self.file_object = io.BytesIO(raw_bytes)
+        self.filename = filename
+        self.filepath = None
+        super().__init__(input_type=InputType.BYTES)
