+++ mindee/input/sources/file_input.py
@@ -1,0 +1,23 @@
+import os
+from typing import BinaryIO
+
+from mindee.input import InputType, LocalInputSource
+
+
+class FileInput(LocalInputSource):
+    """A binary file input."""
+
+    def __init__(self, file: BinaryIO) -> None:
+        """
+        Input document from a Python binary file object.
+
+        Note: the calling function is responsible for closing the file.
+
+        :param file: FileIO object
+        """
+        assert file.name, "File name must be set"
+
+        self.file_object = file
+        self.filename = os.path.basename(file.name)
+        self.filepath = file.name
+        super().__init__(input_type=InputType.FILE)
